Commit 4: método findAll en Service y Controller de User, Loan y Book + método new de Loan + arreglado error de tipo de dato en BookDTO

+++ src/main/java/com/conanthelibrarian/librarymanagementsystem/dto/BookDTO.java
@@ -13,7 +13,7 @@ public class BookDTO {
     private Integer bookId;
     private String title;
     private String author;
-    private Integer isbn;
+    private Long isbn;
     private String genre;
     private Integer availableCopies;
